Determine post attributes in render function
The post attribute variables such as post_id and post_date should be
populated within the render function instead of the render_dir function.
The reason is that when the render function is called directly, the post
attribute variables still needs to be available within the embedded
shell within the post.

Apart from the above change, there is another change in this commit.
A new variable named post_date has been introduced to store the
YYYY-MM-DD part of the post file name. post_create_date is now a synonym
of post_date. This was done so that the variable names post_id and
post_date are consistent with the function names get_post_id and
get_post_date.