Fix test_snip_cache to work on Mac OS X
Consider the following command.

    printf a | sed s/a/b/

This command prints 'b' without a newline on Linux but 'b' followed by
newline on Mac OS X. On Mac OS X, sed adds a newline in the end, if the
input stream did not end with a newline. This difference in behaviour
causes test_snip_cache() to pass on Linux but fail on Mac OS X.

This commit modifies test_snip_cache() such that it does not depend on
the presence or absence of the terminating newline in sed output.