Add out function to print strings safely and neatly
There is no convenient and safe way to print strings in shell. As per
http://pubs.opengroup.org/onlinepubs/009695399/utilities/echo.html the
echo command has the following problem.

> It is not possible to use echo portably across all POSIX systems
> unless both -n (as the first argument) and escape sequences are
> omitted.

The printf command requires the first argument as %s to be used in a
general manner. To overcome these limitations, a function named out is
being added as a utility function. The out function has the following
features.

  - If arguments are specified, it behaves like a safe version of echo.
  - If no arguments are specified, it behaves like the cat command.
  - In both cases, it removes trailing newlines.

The third feature (removing trailing newlines) was added for the
following reason: Due to the way embedded shell code is interpreted and
executed in Oyster, evaluation of many commands produce an additional
newline where there was none in the text with embedded shell code. For
example, the following embedded shell code,

    foo{{ echo bar }}baz
    foo{{ uname -s }}baz

produces the following output.

    foobar
    baz
    fooLinux
    baz

The additional newline introduced after "bar" and "Linux" may not be
desired in some cases. This occurs because many Unix and Linux commands
print a newline in the end and Oyster preserves these newlines while
rendering embedded shell code. In situations where the additional
newlines need to be suppressed, the above embedded shell code may be
written as follows.

    foo{{ out bar }}baz
    foo{{ uname -s | out }}baz

The above code produces the following output when rendered with Oyster.

    foobarbaz
    fooLinuxbaz

Note that the out function removes all trailing newlines, not just the
last one.